was fix project structure
