+++ bin/server.js
@@ -1,7 +1,7 @@
-const app = require('../app')
+const app = require('../app');
 
-const PORT = process.env.PORT || 3000
+const PORT = process.env.PORT || 3000;
 
 app.listen(PORT, () => {
-  console.log(`Server running. Use our API on port: ${PORT}`)
-})
+  console.log(`Server running. Use our API on port: ${PORT}`);
+});
