+++ controllers/contacts/getContact.js
@@ -1,0 +1,12 @@
+const {getContactById} = require('../../model/contacts');
+
+const handlerGetContact = async (req, res, next) => {
+  const id = req.params.id;
+  const contact = await getContactById(id);
+  if (contact) {
+    return res.status(200).json({contact});
+  }
+  return res.status(404).json({message: 'Not found'});
+};
+
+module.exports = handlerGetContact;
