+++ db/contacts.json
@@ -13,8 +13,8 @@
   },
   {
     "id": "3",
-    "name": "Viktor Praid",
-    "email": "mail@mail.com",
+    "name": "Kennedy Lane",
+    "email": "mattis.Cras@nonenimMauris.net",
     "phone": "(542) 451-7038"
   },
   {
@@ -55,26 +55,8 @@
   },
   {
     "id": "10",
-    "name": "Alec Howard",
+    "name": "Bill Murray",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "ab66786b",
-    "name": "Bill Crane",
-    "email": "linch@mail.com",
-    "phone": "(542) 547-8789"
-  },
-  {
-    "id": "e6e38a43",
-    "name": "GarryFrr",
-    "email": "change@mail.com",
-    "phone": "(542) 748-5778"
-  },
-  {
-    "id": "52cee735",
-    "name": "Viktor Praid",
-    "email": "mail@mail.com",
-    "phone": "(542) 787-78-78"
   }
 ]
