+++ routes/api/contacts.js
@@ -2,55 +2,26 @@ const express = require('express');
 const router = express.Router();
 
 const {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-} = require('../../model/');
+  handlerGetAllContacts,
+  handlerGetContact,
+  handlerAddContact,
+  handlerRemoveContact,
+  handlerUpdateContact,
+} = require('../../controllers/contacts');
 
 const {
   addContactValidation,
   changeContactValidation,
-} = require('../../middleware/validationMiddleware');
+} = require('../../middleware/validation/contactValidation');
 
-router.get('/', async (req, res, next) => {
-  const contacts = await listContacts();
-  res.status(200).json({contacts});
-});
+router.get('/', handlerGetAllContacts);
 
-router.get('/:id', async (req, res, next) => {
-  const id = req.params.id;
-  const contact = await getContactById(id);
-  if (contact) {
-    return res.status(200).json({contact});
-  }
-  return res.status(404).json({message: 'Not found'});
-});
+router.get('/:id', handlerGetContact);
 
-router.post('/', addContactValidation, async (req, res, next) => {
-  const body = req.body;
-  const newContact = await addContact(body);
-  res.status(201).json({newContact});
-});
+router.post('/', addContactValidation, handlerAddContact);
 
-router.delete('/:id', async (req, res, next) => {
-  const id = req.params.id;
-  const deleteStatus = await removeContact(id);
-  if (!deleteStatus) {
-    return res.status(404).json({message: 'Not found'});
-  }
-  res.status(200).json({message: 'contact deleted'});
-});
+router.delete('/:id', handlerRemoveContact);
 
-router.put('/:id', changeContactValidation, async (req, res, next) => {
-  const id = req.params.id;
-  const body = req.body;
-  const changeContact = await updateContact(id, body);
-  if (!changeContact) {
-    return res.status(404).json({message: 'Not found'});
-  }
-  res.status(200).json({changeContact});
-});
+router.put('/:id', changeContactValidation, handlerUpdateContact);
 
 module.exports = router;
